backend/requirements.txt
# Core web framework
fastapi>=0.104.0,<0.105.0
uvicorn[standard]>=0.24.0,<0.25.0
python-multipart>=0.0.6

# LangChain ecosystem - let pip resolve compatible versions
langchain>=0.1.0,<0.2.0
langchain-aws>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain-community>=0.0.20,<0.1.0
langgraph>=0.0.40,<0.1.0

# AWS SDK - compatible versions
boto3>=1.34.0
botocore>=1.34.0

# Core dependencies
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.0.0,<3.0.0
python-dotenv>=1.0.0

# Utility libraries
loguru>=0.7.0
Pillow>=10.0.0
requests>=2.31.0
typing-extensions>=4.8.0
orjson>=3.9.0
httpx>=0.25.0
python-dateutil>=2.8.0
aiofiles>=23.0.0

# Development tools (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0

# Production server
gunicorn>=21.0.0

# Security tools
bandit>=1.7.0
safety>=3.0.0